Fine tune object check in Trackable/Event methods.
Upgrade Google test framework to 1.8.0

+++ test/unit/event_virtual/test.cpp
@@ -3,19 +3,16 @@
 #include "test.hpp"
 
 Test::Test()
-    : testing::Test()
-      { }
+    : testing::Test() {}
 
-Test::~Test()
-  { }
+Test::~Test() {}
 
 /*
  *
  */
-TEST_F(Test, connect_method_once)
-{
-  Consumer* obj1 = new Consumer;
-  SubConsumer* obj2 = new SubConsumer;
+TEST_F(Test, connect_method_once) {
+  Consumer *obj1 = new Consumer;
+  SubConsumer *obj2 = new SubConsumer;
 
   cppevent::Event<int> event1;
   event1.Connect(obj1, &Consumer::OnVirtualTest);
@@ -24,7 +21,8 @@ TEST_F(Test, connect_method_once)
   event2.Connect(obj2, &SubConsumer::OnVirtualTest); // this event_connect to the method in sub class
 
   cppevent::Event<int> event3;
-  event3.Connect(dynamic_cast<Consumer*>(obj2), &Consumer::OnVirtualTest); // this still event_connect to the method in sub class
+  event3.Connect(dynamic_cast<Consumer *>(obj2),
+                 &Consumer::OnVirtualTest); // this still event_connect to the method in sub class
 
   event1.Fire(1);
   event2.Fire(2);
@@ -34,20 +32,19 @@ TEST_F(Test, connect_method_once)
 
   delete obj1;
   delete obj2;
-  
+
   ASSERT_TRUE(result);
 }
 
-TEST_F(Test, connect_to_pure_virtual_function)
-{
+TEST_F(Test, connect_to_pure_virtual_function) {
   cppevent::Event<int> event1;
-  AbstractConsumer* obj1 = new Consumer;
+  AbstractConsumer *obj1 = new Consumer;
 
   event1.Connect(obj1, &AbstractConsumer::OnVirtualTest);
 
   event1.Fire(1);
 
-  int result = dynamic_cast<Consumer*>(obj1)->virtualtest_count();
+  size_t result = dynamic_cast<Consumer *>(obj1)->virtualtest_count();
 
   delete obj1;
